% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Simulation code to assess model performance.}
\usage{
simulation(
  setup_level,
  phi_input,
  dist.type,
  n.iter,
  dpw,
  meth,
  convert_weekly,
  NA_type,
  nS,
  year
)
}
\arguments{
\item{setup_level}{character value "daily" or "weekly" dependent on how the data is to be sampled.}

\item{phi_input}{numeric value of phi used to simulate data, at the daily level}

\item{dist.type}{"P", "NB" or "ZIP"}

\item{n.iter}{number of iterations}

\item{dpw}{number of days per week to simulate}

\item{meth}{optimisation method - "Nelder-Mead","SANN" or "L-BFGS-B"}

\item{convert_weekly}{convert from the daily to the weekly format -  "yes" or "no"}

\item{NA_type}{missing values - "match_NA", "prop_NA" or "none"}

\item{nS}{number of sites, if matching raw data then use the length of the output}

\item{year, }{year to simulate}
}
\value{
Vector containing the lifespan estimate for each simulation iteration.
}
\description{
Simulation code to be used to assess biases in model output. Associated with number of days sampled per week or number of sites sampled.
}
\examples{

\donttest{
library(ButterflyLifespan)

# setup simulation parameters
year <- "2018"
setup_level <- "daily"
phi_input <- 0.8
dist.type <- "P" 
n.iter <-250
dpw <- 1
meth <- "Nelder-Mead"
convert_weekly <- "no"
NA_type <- "prop_NA"
nS <- 200

# run simulation
results <- simulation(setup_level,
            phi_input,
		         dist.type, 
		         n.iter,
		         dpw,
		         meth,
		         convert_weekly,
		         NA_type,
		         nS,
		        year)
		        
summary(results$ls_sim) 		        
		        
}

}
